refactor: split cognito.py into cognito_token.py and cognito_admin.py
Separate hot-path (every-request) token verification from cold-path
(rare) Cognito admin operations:

  auth/cognito_token.py (HOT PATH):
    - get_jwks()              — cached JWKS fetch
    - get_signing_key()       — public key lookup by kid
    - verify_token()          — full JWT decode + verify
    - get_email_from_access_token() — email lookup via access token

  auth/cognito_admin.py (COLD PATH):
    - _get_cognito_client()   — cached boto3 client
    - invite_cognito_user()   — create user + send invite
    - delete_cognito_user()   — delete user with fallback

core/cognito.py is now a backward-compat re-export shim so all
existing imports continue to work. It will be removed once all
callers are updated to import from the new locations.
